deploy/RSS-MQ/requirements.txt
feedparser==6.0.8
pika==1.3.1
mysql-connector-python==8.4.0
